Add first bits of logic to app.routes and app.database for path reconstruction on the server.
The search tree method is partially implemented - we can now build a search tree based on a list of observations given, but we don't link it with intersection computation yet.
The logic is figures out - intersection construction needs to be attached to the search tree algorithm so we only perform new intersections for subsets that haven't been seen in the past.

The code is also in need of refactoring.  With Marta's code added to app.database, it will exceed 1000 lines...  this will become difficult to deal with.